==> 5-weather-data/1-parsing-for-weather-data.ipynb <==
# Import

import requests
from bs4 import BeautifulSoup

import pandas as pd
import numpy as np

# Testing

link = "https://www.usclimatedata.com/climate/boston/massachusetts/united-states/usma0046/2012/1"

page = requests.get(link)
page

# creata BeautifulSoup parser
soup = BeautifulSoup(page.content, 'html.parser')

history = soup.find("div", id="history")
# history

left_column = history.find("div", class_="left_column")
#left_column

history_data = left_column.find("table", class_ ="daily_climate_table")
#history_data

# Methods

def method_date(history_data):
    
    date = history_data.findAll(
        "td", 
        class_="align_left daily_climate_table_td_day"
    )

    i = 0 
    array = []
    
    while i < len(date):
        array += date[i]
        i += 1 
    
    return array

def method_temperature_min(history_data):
    
    temperature = history_data.findAll(
        "td", 
        class_="align_right climate_table_data_td temperature_blue"
    )
    
    i = 0 
    array = []
    
    while i < len(temperature):
        array += temperature[i]
        
        i += 1
    
    return array

def method_temperature_max(history_data):
    
    temperature = history_data.findAll(
        "td", 
        class_ = "align_right climate_table_data_td temperature_red "
    )
    
    i = 0 
    array = []
    
    while i < len(temperature1):
        
        array += temperature1[i]
    
        i += 1
    
    return array

def method_precip(history_data):
    
    precip = history_data.findAll(
        "td", 
        class_="align_right climate_table_data_td"
    )

    i = 0
    array = []
    
    while i < len(precip):
        array += precip[i]
        i +=3
    
    return array

def method_snow(history_data):
    
    snow = history_data.findAll(
        "td", 
        class_="align_right climate_table_data_td"
    )
    
    i = 1
    array = []
    
    while i < len(snow):
        array += snow[i]
        i +=3
        
    return array

def method_snow_depth(history_data):
    
    snow_depth = history_data.findAll("td", class_="align_right climate_table_data_td")

    i = 2
    array = []
    while i < len(snow_depth):
        array += snow_depth[i]
            
        i += 3
    
    return array

def create_BS(link):
    
    page = requests.get(link)

    # creata BeautifulSoup parser
    soup = BeautifulSoup(page.content, 'html.parser')
   
    # history
    history = soup.find("div", id="history")
    
    #left_column
    left_column = history.find("div", class_="left_column")

    history_data = left_column.find("table", class_ ="daily_climate_table")

    return history_data

def parse(link):
    
    history_data = create_BS(link)
    
    # THEN
    
    # DATE
    date_array = method_date(history_data)
    
    # temperature min
    temperature_min_array = method_temperature_min(history_data)
    
    # temperature max
    temperature_max_array = method_temperature_max(history_data)
    
    # precip
    precip_array = method_precip(history_data)
    
    # snow
    snow_array = method_snow(history_data)
    
    # snow depth
    snow_depth_array = method_snow_depth(history_data)
    
    
    return date_array, temperature_min_array, temperature_max_array, precip_array, snow_aray, snow_depth_array

# Data

df = pd.DataFrame()

df['date'] = 0
df['temperatureMin'] = 0
df['temperatureMax'] = 0
df['precipitation'] = 0
df['snow'] = 0
df['snow_depth'] = 0


link = "https://www.usclimatedata.com/climate/boston/massachusetts/united-states/usma0046/2012/"

i = 1

while i <=12:
    date_array, temperature_min_array, temperature_max_array, precip_array, snow_aray, snow_depth_array = parse(
    link+str(i)
    )
    
    print(i)
    
    k = 0
    
    while k < len(date_array):
        df = df.append(
            {
                'date': date_array[k],
                'temperatureMin': temperature_array[k],
                'temperatureMax': temperature1_array[k],
                'precipitation': precip_array[k],
                'snow': snow_aray[k],
                'snow_depth': snow_depth_array[k]  
            },
            ignore_index=True          
        )
        k += 1
        
    i += 1

link = "https://www.usclimatedata.com/climate/boston/massachusetts/united-states/usma0046/2013/"

i = 1

while i <=12:
    date_array, temperature_min_array, temperature_max_array, precip_array, snow_aray, snow_depth_array = parse(
    link+str(i)
    )
    
    print(i)
    
    k = 0
    
    while k < len(date_array):
        df = df.append(
            {
                'date': date_array[k],
                'temperatureMin': temperature_array[k],
                'temperatureMax': temperature1_array[k],
                'precipitation': precip_array[k],
                'snow': snow_aray[k],
                'snow_depth': snow_depth_array[k]  
            },
            ignore_index=True          
        )
        k += 1
        
    i += 1

link = "https://www.usclimatedata.com/climate/boston/massachusetts/united-states/usma0046/2014/"

i = 1

while i <=12:
    date_array, temperature_min_array, temperature_max_array, precip_array, snow_aray, snow_depth_array = parse(
    link+str(i)
    )
    
    print(i)
    
    k = 0
    
    while k < len(date_array):
        df = df.append(
            {
                'date': date_array[k],
                'temperatureMin': temperature_array[k],
                'temperatureMax': temperature1_array[k],
                'precipitation': precip_array[k],
                'snow': snow_aray[k],
                'snow_depth': snow_depth_array[k]  
            },
            ignore_index=True          
        )
        k += 1
        
    i += 1

link = "https://www.usclimatedata.com/climate/boston/massachusetts/united-states/usma0046/2015/"

i = 1

while i <=12:
    date_array, temperature_min_array, temperature_max_array, precip_array, snow_aray, snow_depth_array = parse(
    link+str(i)
    )
    
    print(i)
    
    k = 0
    
    while k < len(date_array):
        df = df.append(
            {
                'date': date_array[k],
                'temperatureMin': temperature_array[k],
                'temperatureMax': temperature1_array[k],
                'precipitation': precip_array[k],
                'snow': snow_aray[k],
                'snow_depth': snow_depth_array[k]  
            },
            ignore_index=True          
        )
        k += 1
        
    i += 1

link = "https://www.usclimatedata.com/climate/boston/massachusetts/united-states/usma0046/2016/"

i = 1

while i <=12:
    date_array, temperature_min_array, temperature_max_array, precip_array, snow_aray, snow_depth_array = parse(
    link+str(i)
    )
    
    print(i)
    
    k = 0
    
    while k < len(date_array):
        df = df.append(
            {
                'date': date_array[k],
                'temperatureMin': temperature_array[k],
                'temperatureMax': temperature1_array[k],
                'precipitation': precip_array[k],
                'snow': snow_aray[k],
                'snow_depth': snow_depth_array[k]  
            },
            ignore_index=True          
        )
        k += 1
        
    i += 1

link = "https://www.usclimatedata.com/climate/boston/massachusetts/united-states/usma0046/2017/"

i = 1

while i <=12:
    date_array, temperature_min_array, temperature_max_array, precip_array, snow_aray, snow_depth_array = parse(
    link+str(i)
    )
    
    print(i)
    
    k = 0
    
    while k < len(date_array):
        df = df.append(
            {
                'date': date_array[k],
                'temperatureMin': temperature_array[k],
                'temperatureMax': temperature1_array[k],
                'precipitation': precip_array[k],
                'snow': snow_aray[k],
                'snow_depth': snow_depth_array[k]  
            },
            ignore_index=True          
        )
        k += 1
        
    i += 1

link = "https://www.usclimatedata.com/climate/boston/massachusetts/united-states/usma0046/2018/"

i = 1

while i <=12:
    date_array, temperature_min_array, temperature_max_array, precip_array, snow_aray, snow_depth_array = parse(
    link+str(i)
    )
    
    print(i)
    
    k = 0
    
    while k < len(date_array):
        df = df.append(
            {
                'date': date_array[k],
                'temperatureMin': temperature_array[k],
                'temperatureMax': temperature1_array[k],
                'precipitation': precip_array[k],
                'snow': snow_aray[k],
                'snow_depth': snow_depth_array[k]  
            },
            ignore_index=True          
        )
        k += 1
        
    i += 1

df

df.to_csv('total-weather.csv')
