==> 5-weather-data/2-preparation-weather-data-for-merging.ipynb <==
# Import

import numpy as np
import pandas as pd

# Read data

df = pd.read_csv('total-weather.csv')

df.head()

# Preparation

df = df.drop('Unnamed: 0', 1)

df.date = pd.to_datetime(df.date)

df['year'] = df['date'].apply(lambda x: x.year)

df.year.unique()

df['month'] = df['date'].apply(lambda x: x.month)

df.month.unique()

df['day'] = df['date'].apply(lambda x: x.day)

df.day.unique()

df.head()

df.temperatureMin.describe()

df.temperatureMax.describe()

df.snow.unique()

df.snow_depth.unique()

df.precipitation = pd.to_numeric(df.precipitation, errors='coerce')

df.precipitation.describe()

df.head()

df = df.drop('snow_depth',1)

df.to_csv('total-weather-for-merge.csv')
