==> 5-weather-data/3-merge-offenses-data-with-weather.ipynb <==
# Import

import pandas as pd

# Read data

w = pd.read_csv('total-weather-for-merge.csv')

w.head()

w = w.drop('Unnamed: 0', 1)

w.snow = pd.to_numeric(w.snow, errors='coerce')

w.describe()

c = pd.read_csv('crime-boston_1.csv')

c.head()

c = c.drop('Unnamed: 0',1)

w.columns

c.OCCURRED_ON_DATE = pd.to_datetime(c.OCCURRED_ON_DATE)

c['DAY'] = c.OCCURRED_ON_DATE.dt.day

c['temperatureMin'] = 0
c['temperatureMax'] = 0
c['precipitation'] = 0
c['snow'] = 0

# Merging

year = 2012
m = 7

while m <=12:
    
    d = 1
    
    while d <= len(w.day.loc[(w.year == year) & (w.month == m)].unique()):
        
        # temperatureMin
        tmin = w.temperatureMin.loc[
            (w.year == year) & (w.month == m) & (w.day == d)
        ]
        
        c.temperatureMin.loc[
            (c.YEAR == year) & (c.MONTH == m) & (c.DAY == d)
        ] = float(tmin)
        
        # temperatureMin
        tmax = w.temperatureMax.loc[
            (w.year == year) & (w.month == m) & (w.day == d)
        ]
        
        c.temperatureMax.loc[
            (c.YEAR == year) & (c.MONTH == m) & (c.DAY == d)
        ] = float(tmax)
        
        
        # precipitation
        p = w.precipitation.loc[
            (w.year == year) & (w.month == m) & (w.day == d)
        ]
        
        c.precipitation.loc[
            (c.YEAR == year) & (c.MONTH == m) & (c.DAY == d)
        ] = float(p)
        
        # snow
        sn = w.snow.loc[
            (w.year == year) & (w.month == m) & (w.day == d)
        ]
        
        c.snow.loc[
            (c.YEAR == year) & (c.MONTH == m) & (c.DAY == d)
        ] = float(sn)
        
        d +=1
    
    m += 1

year = 2013
m = 1

while m <=12:
    
    d = 1
    
    while d <= len(w.day.loc[(w.year == year) & (w.month == m)].unique()):
        
        # temperatureMin
        tmin = w.temperatureMin.loc[
            (w.year == year) & (w.month == m) & (w.day == d)
        ]
        
        c.temperatureMin.loc[
            (c.YEAR == year) & (c.MONTH == m) & (c.DAY == d)
        ] = float(tmin)
        
        # temperatureMin
        tmax = w.temperatureMax.loc[
            (w.year == year) & (w.month == m) & (w.day == d)
        ]
        
        c.temperatureMax.loc[
            (c.YEAR == year) & (c.MONTH == m) & (c.DAY == d)
        ] = float(tmax)
        
        
        # precipitation
        p = w.precipitation.loc[
            (w.year == year) & (w.month == m) & (w.day == d)
        ]
        
        c.precipitation.loc[
            (c.YEAR == year) & (c.MONTH == m) & (c.DAY == d)
        ] = float(p)
        
        # snow
        sn = w.snow.loc[
            (w.year == year) & (w.month == m) & (w.day == d)
        ]
        
        c.snow.loc[
            (c.YEAR == year) & (c.MONTH == m) & (c.DAY == d)
        ] = float(sn)
        
        d +=1
    
    m += 1

year = 2014
m = 1

while m <=12:
    
    d = 1
    
    while d <= len(w.day.loc[(w.year == year) & (w.month == m)].unique()):
        
        # temperatureMin
        tmin = w.temperatureMin.loc[
            (w.year == year) & (w.month == m) & (w.day == d)
        ]
        
        c.temperatureMin.loc[
            (c.YEAR == year) & (c.MONTH == m) & (c.DAY == d)
        ] = float(tmin)
        
        # temperatureMin
        tmax = w.temperatureMax.loc[
            (w.year == year) & (w.month == m) & (w.day == d)
        ]
        
        c.temperatureMax.loc[
            (c.YEAR == year) & (c.MONTH == m) & (c.DAY == d)
        ] = float(tmax)
        
        
        # precipitation
        p = w.precipitation.loc[
            (w.year == year) & (w.month == m) & (w.day == d)
        ]
        
        c.precipitation.loc[
            (c.YEAR == year) & (c.MONTH == m) & (c.DAY == d)
        ] = float(p)
        
        # snow
        sn = w.snow.loc[
            (w.year == year) & (w.month == m) & (w.day == d)
        ]
        
        c.snow.loc[
            (c.YEAR == year) & (c.MONTH == m) & (c.DAY == d)
        ] = float(sn)
        
        d +=1
    
    m += 1

year = 2015
m = 1

while m <=12:
    
    d = 1
    
    while d <= len(w.day.loc[(w.year == year) & (w.month == m)].unique()):
        
        # temperatureMin
        tmin = w.temperatureMin.loc[
            (w.year == year) & (w.month == m) & (w.day == d)
        ]
        
        c.temperatureMin.loc[
            (c.YEAR == year) & (c.MONTH == m) & (c.DAY == d)
        ] = float(tmin)
        
        # temperatureMin
        tmax = w.temperatureMax.loc[
            (w.year == year) & (w.month == m) & (w.day == d)
        ]
        
        c.temperatureMax.loc[
            (c.YEAR == year) & (c.MONTH == m) & (c.DAY == d)
        ] = float(tmax)
        
        
        # precipitation
        p = w.precipitation.loc[
            (w.year == year) & (w.month == m) & (w.day == d)
        ]
        
        c.precipitation.loc[
            (c.YEAR == year) & (c.MONTH == m) & (c.DAY == d)
        ] = float(p)
        
        # snow
        sn = w.snow.loc[
            (w.year == year) & (w.month == m) & (w.day == d)
        ]
        
        c.snow.loc[
            (c.YEAR == year) & (c.MONTH == m) & (c.DAY == d)
        ] = float(sn)
        
        d +=1
    
    m += 1

year = 2016
m = 1

while m <=12:
    
    d = 1
    
    while d <= len(w.day.loc[(w.year == year) & (w.month == m)].unique()):
        
        # temperatureMin
        tmin = w.temperatureMin.loc[
            (w.year == year) & (w.month == m) & (w.day == d)
        ]
        
        c.temperatureMin.loc[
            (c.YEAR == year) & (c.MONTH == m) & (c.DAY == d)
        ] = float(tmin)
        
        # temperatureMin
        tmax = w.temperatureMax.loc[
            (w.year == year) & (w.month == m) & (w.day == d)
        ]
        
        c.temperatureMax.loc[
            (c.YEAR == year) & (c.MONTH == m) & (c.DAY == d)
        ] = float(tmax)
        
        
        # precipitation
        p = w.precipitation.loc[
            (w.year == year) & (w.month == m) & (w.day == d)
        ]
        
        c.precipitation.loc[
            (c.YEAR == year) & (c.MONTH == m) & (c.DAY == d)
        ] = float(p)
        
        # snow
        sn = w.snow.loc[
            (w.year == year) & (w.month == m) & (w.day == d)
        ]
        
        c.snow.loc[
            (c.YEAR == year) & (c.MONTH == m) & (c.DAY == d)
        ] = float(sn)
        
        d +=1
    
    m += 1

year = 2017
m = 1

while m <=12:
    
    d = 1
    
    while d <= len(w.day.loc[(w.year == year) & (w.month == m)].unique()):
        
        # temperatureMin
        tmin = w.temperatureMin.loc[
            (w.year == year) & (w.month == m) & (w.day == d)
        ]
        
        c.temperatureMin.loc[
            (c.YEAR == year) & (c.MONTH == m) & (c.DAY == d)
        ] = float(tmin)
        
        # temperatureMin
        tmax = w.temperatureMax.loc[
            (w.year == year) & (w.month == m) & (w.day == d)
        ]
        
        c.temperatureMax.loc[
            (c.YEAR == year) & (c.MONTH == m) & (c.DAY == d)
        ] = float(tmax)
        
        
        # precipitation
        p = w.precipitation.loc[
            (w.year == year) & (w.month == m) & (w.day == d)
        ]
        
        c.precipitation.loc[
            (c.YEAR == year) & (c.MONTH == m) & (c.DAY == d)
        ] = float(p)
        
        # snow
        sn = w.snow.loc[
            (w.year == year) & (w.month == m) & (w.day == d)
        ]
        
        c.snow.loc[
            (c.YEAR == year) & (c.MONTH == m) & (c.DAY == d)
        ] = float(sn)
        
        d +=1
    
    m += 1

year = 2018
m = 1

while m <=12:
    
    d = 1
    
    while d <= len(w.day.loc[(w.year == year) & (w.month == m)].unique()):
        
        # temperatureMin
        tmin = w.temperatureMin.loc[
            (w.year == year) & (w.month == m) & (w.day == d)
        ]
        
        c.temperatureMin.loc[
            (c.YEAR == year) & (c.MONTH == m) & (c.DAY == d)
        ] = float(tmin)
        
        # temperatureMin
        tmax = w.temperatureMax.loc[
            (w.year == year) & (w.month == m) & (w.day == d)
        ]
        
        c.temperatureMax.loc[
            (c.YEAR == year) & (c.MONTH == m) & (c.DAY == d)
        ] = float(tmax)
        
        
        # precipitation
        p = w.precipitation.loc[
            (w.year == year) & (w.month == m) & (w.day == d)
        ]
        
        c.precipitation.loc[
            (c.YEAR == year) & (c.MONTH == m) & (c.DAY == d)
        ] = float(p)
        
        # snow
        sn = w.snow.loc[
            (w.year == year) & (w.month == m) & (w.day == d)
        ]
        
        c.snow.loc[
            (c.YEAR == year) & (c.MONTH == m) & (c.DAY == d)
        ] = float(sn)
        
        d +=1
    
    m += 1

c[['temperatureMin','temperatureMax','precipitation','snow']].describe()

c['temperatureDifference'] = c['temperatureMax'] - c['temperatureMin']

c['temperatureDifference'].describe()

c[['temperatureDifference','temperatureMax','temperatureMin']].head()

c.shape

c.isnull().sum()

c.to_csv('finish-crime-weather.csv')
