==> 5-weather-data/4-research-for-Offense_code_groups.ipynb <==
# 1. Import

import numpy as np
import pandas as pd

# 2. Read data

df = pd.read_csv('crimes.csv')

df.columns

df = df.drop('Unnamed: 0',1)

df1 = pd.read_csv('crime-incident-reports-july-2012---august-2015-source-legacy-system.csv')

df1.columns

df2 = pd.read_csv('crimeincidentreports.csv')

df2.columns

# 3. Research -> OFFENSE_CODE_GROUP

print(len(df.OFFENSE_CODE_GROUP.unique()))

df.OFFENSE_CODE_GROUP.unique()

df.OFFENSE_CODE_GROUP.value_counts()

### 3.1. Violations

###### First

I research description and other values for this offense code grop (Violations).

# for first data set (july-2012---august-2015)

# VAL => Violations

df1.MAIN_CRIMECODE.loc[df1.INCIDENT_TYPE_DESCRIPTION =='VAL'].unique()

# for first data set (july-2012---august-2015)

# VAL => Violations

df1.Shooting.loc[df1.INCIDENT_TYPE_DESCRIPTION =='VAL'].unique()

# for first data set (july-2012---august-2015)

# VAL => Violations

df1.Shooting.loc[df1.INCIDENT_TYPE_DESCRIPTION =='VAL'].unique()

# for first data set (july-2012---august-2015)

# VAL => Violations

df1.UCRPART.loc[df1.INCIDENT_TYPE_DESCRIPTION =='VAL'].unique()

# for second data set (august-2015---todate)

df2.OFFENSE_CODE.loc[df2.OFFENSE_CODE_GROUP=='Violations'].unique()

# for second data set (august-2015---todate)

df2.OFFENSE_DESCRIPTION.loc[df2.OFFENSE_CODE_GROUP=='Violations'].unique()

# for second data set (august-2015---todate)

df2.SHOOTING.loc[df2.OFFENSE_CODE_GROUP=='Violations'].unique()

# for second data set (august-2015---todate)

df2.UCR_PART.loc[df2.OFFENSE_CODE_GROUP=='Violations'].unique()

###### Second

Definitions:

A **violation** of law is any act (or, less commonly, failure to act) that fails to abide by existing law. Violations generally may include both crimes and civil wrongs. Some acts, such as fraud, can violate both civil and criminal laws.

*Examples of violations of law include:*

- Moving violation, any violation of law committed by a driver while the vehicle is in motion
- Parking violation, parking a motor vehicle in a restricted place or an unauthorized manner
- Willful violation, in U.S. law an act with intentional disregard for a regulation, statute and policy
- Infraction, in United States law, minor or petty offenses that do not require jury trial
- Infringement, various violations of laws or rights, usually used in the context of intellectual property

Wikipedia: https://en.wikipedia.org/wiki/Violation_of_law 

###### Hypothesis for input data

- traffic

### 3.2. Larceny

# Incident type description for larceny - > OTHER LARCENY / Larceny

df1.MAIN_CRIMECODE.loc[df1.INCIDENT_TYPE_DESCRIPTION =='OTHER LARCENY'].unique()

df1.MAIN_CRIMECODE.loc[df1.INCIDENT_TYPE_DESCRIPTION =='Larceny'].unique()

df1.Shooting.loc[df1.INCIDENT_TYPE_DESCRIPTION =='Larceny'].unique()

df1.Shooting.loc[df1.INCIDENT_TYPE_DESCRIPTION =='OTHER LARCENY'].unique()

df1.UCRPART.loc[df1.INCIDENT_TYPE_DESCRIPTION =='Larceny'].unique()

df1.UCRPART.loc[df1.INCIDENT_TYPE_DESCRIPTION =='OTHER LARCENY'].unique()

df2.OFFENSE_CODE.loc[df2.OFFENSE_CODE_GROUP == 'Larceny'].unique()

df2.OFFENSE_DESCRIPTION.loc[df2.OFFENSE_CODE_GROUP == 'Larceny'].unique()

df2.OFFENSE_DESCRIPTION.loc[df2.OFFENSE_CODE_GROUP == 'Larceny'].value_counts()

df2.UCR_PART.loc[df2.OFFENSE_CODE_GROUP == 'Larceny'].unique()

df2.SHOOTING.loc[df2.OFFENSE_CODE_GROUP == 'Larceny'].unique()

### 3.3. Simple Assault 

df2.UCR_PART.loc[df2.OFFENSE_CODE_GROUP=='Simple Assault'].unique()

df2.OFFENSE_DESCRIPTION.loc[df2.OFFENSE_CODE_GROUP=='Simple Assault'].unique()

df2.OFFENSE_DESCRIPTION.loc[df2.OFFENSE_CODE_GROUP=='Simple Assault'].value_counts()

### 3.4. Medical Assistance

df2.UCR_PART.loc[df2.OFFENSE_CODE_GROUP=='Medical Assistance'].unique()

df2.OFFENSE_DESCRIPTION.loc[df2.OFFENSE_CODE_GROUP=='Medical Assistance'].unique()

df2.OFFENSE_DESCRIPTION.loc[df2.OFFENSE_CODE_GROUP=='Medical Assistance'].value_counts()

### 3.5. Motor Vehicle Accident Response

df2.UCR_PART.loc[df2.OFFENSE_CODE_GROUP=='Motor Vehicle Accident Response'].unique()

df2.OFFENSE_DESCRIPTION.loc[df2.OFFENSE_CODE_GROUP=='Motor Vehicle Accident Response'].unique()

df2.OFFENSE_DESCRIPTION.loc[df2.OFFENSE_CODE_GROUP=='Motor Vehicle Accident Response'].value_counts()

### 3.6. Vandalism

df2.UCR_PART.loc[df2.OFFENSE_CODE_GROUP=='Vandalism'].unique()

df2.OFFENSE_DESCRIPTION.loc[df2.OFFENSE_CODE_GROUP=='Vandalism'].unique()

### 3.7. Investigate Person

df2.UCR_PART.loc[df2.OFFENSE_CODE_GROUP=='Investigate Person'].unique()

df2.OFFENSE_DESCRIPTION.loc[df2.OFFENSE_CODE_GROUP=='Investigate Person'].unique()

### 3.8. Larceny From Motor Vehicle

df2.UCR_PART.loc[df2.OFFENSE_CODE_GROUP == 'Larceny From Motor Vehicle'].unique()

df2.OFFENSE_CODE_GROUP.loc[df2.OFFENSE_CODE_GROUP == 'Larceny From Motor Vehicle'].unique()

### 3.9. Drug Violation

df2.UCR_PART.loc[df2.OFFENSE_CODE_GROUP == 'Drug Violation'].unique()

df2.OFFENSE_DESCRIPTION.loc[df2.OFFENSE_CODE_GROUP == 'Drug Violation'].unique()

df2.OFFENSE_DESCRIPTION.loc[df2.OFFENSE_CODE_GROUP == 'Drug Violation'].value_counts()

### 3.10. Property Lost

df2.UCR_PART.loc[df2.OFFENSE_CODE_GROUP == 'Property Lost'].unique()

df2.OFFENSE_DESCRIPTION.loc[df2.OFFENSE_CODE_GROUP == 'Property Lost'].unique()

### 3.11. Fraud

df2.UCR_PART.loc[df2.OFFENSE_CODE_GROUP == 'Fraud'].unique()

df2.OFFENSE_DESCRIPTION.loc[df2.OFFENSE_CODE_GROUP == 'Fraud'].unique()

### 3.12. Towed

df2.UCR_PART.loc[df2.OFFENSE_CODE_GROUP == 'Towed'].unique()

df2.OFFENSE_DESCRIPTION.loc[df2.OFFENSE_CODE_GROUP == 'Towed'].unique()

df.OFFENSE_CODE_GROUP.value_counts()
